reformatting app to be more a modular package
